Use bigint columns for GitHub App ids
CI did not expose usable logs, but the new GitHub App schema had MySQL-facing 32-bit integer columns for GitHub identifiers. Those ids can exceed signed integer range, which can break installation sync or callback persistence under MySQL.

Changed the new AppSetting and Installation identifier columns to `bigint` and added model specs covering large GitHub App, installation, and account ids. Local lint, security scans, importmap audit, targeted specs, and full RSpec passed.

+++ db/migrate/20260511140000_add_github_app_credentials_to_app_settings.rb
@@ -1,6 +1,6 @@
 class AddGithubAppCredentialsToAppSettings < ActiveRecord::Migration[8.1]
   def change
-    add_column :app_settings, :github_app_id, :integer
+    add_column :app_settings, :github_app_id, :bigint
     add_column :app_settings, :github_app_slug, :string
     add_column :app_settings, :github_app_private_key_pem, :text
     add_column :app_settings, :github_app_webhook_secret, :text
